Added VHDL_SPI project and updated SPI on Tiva
Updated the SPI implementation on the Tiva board to accommodate the SPI implementation on the FPGA board. Likewise, initial SPI implementation has been added for the FPGA board. In this, there is also code for the FPGA board in master mode, but this has mainly been used for testing purposes. For the project, the S_SPI_TOPMODULE should suffice.